+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 販売者のmakesOfferに適用取引期間を追加
+- PayActionを最適化
+- PayActionにlocationを追加
 
 ### Deprecated